% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ab_basic.R
\name{boot_ab_basic}
\alias{boot_ab_basic}
\title{Bootstrap Basic CI for A/B conversion rates}
\usage{
boot_ab_basic(nA, yA, nB, yB, B = 2000, conf = 0.95)
}
\arguments{
\item{nA}{Sample size of group A.}

\item{yA}{Successes in group A.}

\item{nB}{Sample size of group B.}

\item{yB}{Successes in group B.}

\item{B}{Number of bootstrap replicates.}

\item{conf}{Confidence level (default 0.95).}
}
\value{
A list with:
\itemize{
  \item pA conversion rate in A
  \item pB conversion rate in B
  \item diff difference pB - pA
  \item ci basic bootstrap confidence interval
}
}
\description{
This function computes the bootstrap *basic* confidence interval
for the difference in conversion rates between A and B.
}
\examples{
boot_ab_basic(1000, 70, 1000, 95, B = 500)
}
