% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ab_perc.R
\name{boot_ab_perc}
\alias{boot_ab_perc}
\title{Bootstrap percentile CI for A/B conversion rates}
\usage{
boot_ab_perc(nA, yA, nB, yB, B = 2000, conf = 0.95)
}
\arguments{
\item{nA}{Integer. Sample size of group A.}

\item{yA}{Integer. Number of successes in group A.}

\item{nB}{Integer. Sample size of group B.}

\item{yB}{Integer. Number of successes in group B.}

\item{B}{Integer. Number of bootstrap replications.}

\item{conf}{Confidence level (default 0.95)}
}
\value{
A list containing:
\itemize{
  \item pA Estimated conversion rate in A
  \item pB Estimated conversion rate in B
  \item diff Difference (pB - pA)
  \item ci Percentile bootstrap CI
}
}
\description{
This function performs a bootstrap percentile confidence interval
for the difference in conversion rates between group A and group B.
}
\examples{
boot_ab_perc(1000, 70, 1000, 95, B = 500)
}
