% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ab_t.R
\name{boot_ab_t}
\alias{boot_ab_t}
\title{Bootstrap-t CI for A/B conversion rates}
\usage{
boot_ab_t(nA, yA, nB, yB, B = 2000, conf = 0.95)
}
\arguments{
\item{nA}{Sample size of group A.}

\item{yA}{Successes in group A.}

\item{nB}{Sample size of group B.}

\item{yB}{Successes in group B.}

\item{B}{Number of bootstrap replicates.}

\item{conf}{Confidence level (default 0.95).}
}
\value{
A list containing:
\itemize{
  \item pA conversion rate for group A
  \item pB conversion rate for group B
  \item diff difference pB - pA
  \item ci bootstrap-t confidence interval
}
}
\description{
This function computes the *bootstrap-t* confidence interval
for the difference in conversion rates between A and B.
}
\examples{
boot_ab_t(1000, 70, 1000, 95, B = 500)
}
